packages/alisa-data/pyproject.toml
[project]
name = "alisa-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.5.0",
    "aiomysql>=0.3.2",
    "asyncpg>=0.31.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pymysql>=1.1.2",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.48",
]
[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"
